misc: fix MSVC compiler warnings

+++ examples/cpu/stream.cpp
@@ -17,11 +17,11 @@
 
 const unsigned DIMX = 640;
 const unsigned DIMY = 480;
-const float PI = 3.14159265359;
+const float PI = 3.14159265359f;
 const float MINIMUM = 1.0f;
 const float MAXIMUM = 20.f;
 const float STEP    = 2.0f;
-const int NELEMS    = (MAXIMUM-MINIMUM+1)/STEP;
+const int NELEMS    = (int)((MAXIMUM-MINIMUM+1)/STEP);
 
 using namespace std;
